Huge interactables refactor
Biggest tweak in platforms and switches:

Switches now only just send a signal, and movable platforms listen for it.

The previous logic made buttons and switches handle movable platforms positions and such, but now is not the case anymore.
Movable platforms now inherit from a switch listener superclass, and allow for position, rotation and scale to be animated too.

Also added DOTween plugin.